+++ src/core/user/domain/domainServices/user.domainServices.ts
@@ -1,16 +1,16 @@
 import { Injectable } from '@nestjs/common';
-import { UserRepositories } from '../repositories/user.repositories';
+import { UserRepositories } from '../../repositories/user.repositories';
 import { registerEmailTemplate } from 'src/email-templates/register';
 import { MailOptions } from 'nodemailer/lib/smtp-transport';
-import { UserSecurity } from '../security/user.security';
-import { REGISTER_REQUEST_RETRY_DAY, USER_S3_BUCKET } from '../user.config';
-import { DateTime } from 'luxon';
+import { UserSecurity } from '../../security/user.security';
+import { USER_S3_BUCKET } from '../../user.config';
 import crypto from 'bcrypt';
-import { FORGOT_PASS_EXT, LOGIN_EXT } from '../user.config';
+import { FORGOT_PASS_EXT, LOGIN_EXT } from '../../user.config';
 import { S3Service } from 'src/Infrastructure/s3/s3.service';
 import { User } from '@prisma/client';
 import fs from 'fs';
 import path from 'path';
+import { approveRegisterRequest, registerRequest, register } from './register.domainServices';
 
 
 @Injectable()
@@ -30,51 +30,7 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   }> {
-    try {
-      const registerRequest = await this.userRepositories.findRegisterRequest(email);
-
-      if (!registerRequest) {
-        await this.userRepositories.createRegisterRequest({ email });
-      } else {
-        const now = DateTime.now();
-        const lastRequest = DateTime.fromJSDate(registerRequest.updatedAt);
-        const dayDiff = now.diff(lastRequest, 'days').days;
-
-        if (dayDiff < REGISTER_REQUEST_RETRY_DAY) {
-          return {
-            isError: true,
-            message: `Your request being processed. Please wait for ${(REGISTER_REQUEST_RETRY_DAY - dayDiff).toFixed()} days before requesting again`,
-          };
-        } else {
-          await this.userRepositories.findAndUpdateRegisterRequest(email);
-        }
-      }
-
-      const emailContent = registerEmailTemplate(
-        'Thank you for registering with us!',
-        'You will receive a separate email with a registration link once your email is approved. Please allow up to 2 business days for processing.',
-      );
-
-      const mailOptions = {
-        from: 'admin@wliafdew.dev',
-        to: email,
-        subject: 'Register request',
-        html: emailContent,
-      };
-
-      return {
-        isError: false,
-        message: 'Register request created successfully',
-        mailOptions,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error creating register request',
-      };
-    }
+    return await registerRequest(email, this.userRepositories)
   }
 
   approveRegisterRequest(email: string): {
@@ -83,54 +39,7 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   } {
-    try {
-      const signToken = this.userSecurityServices.signJWT(
-        {
-          email,
-        },
-        '3d',
-        'REGISTER_REQUEST',
-        false
-      );
-
-      if (!signToken) {
-        throw new Error('Error signing token');
-      }
-
-      const emailContent = registerEmailTemplate(
-        'Thank you for patience!',
-        ` 
-                Your email has been approved. Please click the link below to complete your registration. This link will expire in 3 days. 
-                If you did not request this, please ignore this email.
-                <br/>
-                <a href="${process.env.FRONTEND_URL}/register?p=${signToken}"
-                style="color: rgb(0, 141, 163); --darkreader-inline-color: #5ae9ff; margin-top: 10px;"
-                target="_blank"
-                data-saferedirecturl="">Register Link!
-               </a> 
-            `,
-      );
-
-      const mailOptions = {
-        from: 'admin@wliafdew.dev',
-        to: email,
-        subject: 'Register request',
-        html: emailContent,
-      };
-
-      return {
-        isError: false,
-        message: 'Register request approved',
-        mailOptions,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error approving register request',
-      };
-    }
+    return approveRegisterRequest(email, this.userSecurityServices);
   }
 
   async register(data: {
@@ -144,39 +53,7 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   }> {
-    try {
-      const hashedPassword = await crypto.hash(data.password, 10);
-      await this.userRepositories.createUser({
-        email: data.email,
-        password: hashedPassword,
-        name: data.name,
-      });
-
-      const emailContent = registerEmailTemplate(
-        'Welcome to our platform!',
-        'You have successfully registered with us. You can now login to your account.',
-      );
-
-      const mailOptions = {
-        from: 'admin@wliafdew.dev',
-        to: data.email,
-        subject: 'Register request',
-        html: emailContent,
-      };
-
-      return {
-        isError: false,
-        message: 'User created successfully',
-        mailOptions,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error create user',
-      };
-    }
+    return  await register(data, this.userRepositories)
   }
 
   async login(credentials: { email: string; password: string }): Promise<{
@@ -787,9 +664,6 @@ export class UserDomainServices {
           message: 'Invalid access token',
         }
       }
-
-
-
 
 
       //perform upload to s3
