+++ src/core/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserService } from './services/user.service';
 import { UserController } from 'src/presentation/controllers/user.controller';
 import { NodemailerService } from 'src/Infrastructure/nodemailer/nodemailer.service';
 import { UserRepositories } from './repositories/user.repositories';
-import { UserDomainServices } from './domain/user.domainServices';
+import { UserDomainServices } from './domain/domainServices/user.domainServices';
 import { PrismaService } from 'src/Infrastructure/prisma-client/prisma-client.service';
 import { RegisterRequestGuard } from './guards/registerRequest.guard';
 import { UserSecurity } from './security/user.security';
